TST #725 Will CI run this now?
